azure-jwt-validator/requirements.txt
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
